// frontend/src/api/stockHealthEval.ts
export interface StockHealthEval {
  id: number;
  exchange: string;
  ticker: string;
  currency: string;
  favorite: boolean;
  industry: string;
  nextEarningsDate: Date | undefined;
  latestPrice: number | undefined;
  positiveOperatingCashFlow: boolean | undefined;
  operatingCashFlowHealth: boolean | undefined;
  improvingGrossMargin: boolean | undefined;
  lowCustomerAcquisition: boolean | undefined;
  improvingCustomerAcquisition: boolean | undefined;
  lowInnovation: boolean | undefined;
  improvingInnovation: boolean | undefined;
  lowInterestBurden: boolean | undefined;
  healthyNetCash: boolean | undefined;
  positiveRetainedEarnings: boolean | undefined;
  improvingRetainedEarnings: boolean | undefined;
  noShareDilution: boolean | undefined;
  imrprovingShareDilution: boolean | undefined;
  latestRevenue: number | undefined;
  revenueNextYear: number | undefined;
  revenueGrowth1y: number | undefined;
  revenueGrowthMultiYear: number | undefined;
  revenueGrowthNextYear: number | undefined;
  grossProfitGrowth1y: number | undefined;
  grossProfitGrowthMultiYear: number | undefined;
  latestGrossMargin: number | undefined;
  latestOperatingMargin: number | undefined;
  latestNetMargin: number | undefined;
  latestEps: number | undefined;
  latestOperatingCashFlow: number | undefined;
  latestFreegCashFlow: number | undefined;
  theoreticalNetMargin: number | undefined;
  theoreticalNetIncome: number | undefined;
  theoreticalNetIncomeNextYear: number | undefined;
  optimizedNetMargin: boolean | undefined;
  priceCurrentRevenueGrowth: number | undefined;
  priceCurrentGpGrowth: number | undefined;
  priceNextYearRevenueGrowth: number | undefined;
  priceMultiYearRevenueGrowth: number | undefined;
  priceMultiYearGpGrowth: number | undefined;
  lastUpdated: Date | undefined;
}